requirements.txt
# AI Governance Starter Kit - Runtime Dependencies
#
# These are the core dependencies needed to run models and governance tools.
# Install with: pip install -r requirements.txt

# YAML processing for model registry
pyyaml>=6.0.1

# Data processing and ML
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Model serialization
joblib>=1.3.0

# For monitoring and drift detection
scipy>=1.11.0

# Logging and utilities
python-dateutil>=2.8.2
